Add public network and quintupleo template
This should have been two commits, but apparently I forgot to push
the button after the first one.

Adding a public network allows us to simulate a more realistic
environment where the provisioning network and the public API
network are separate.

The QuintupleO template wraps the existing virtual-baremetal.yaml
and does some extra configuration suited to a TripleO deployment
model.  I made it a separate template in an attempt to decouple
OVB and QuintupleO since the former doesn't necessarily need the
latter.