Allow a single BMC to manage multiple BMs
This change creates a single BMC server and uses that to manage multiple
baremetal servers.

It does this by using a naming convention on the BM and BMC neutron
ports and using that to match a BMC port to a BM nova server.

In the BMC boot script a separate systemd unit is created per network
interface, and the IP address of that interface is used to bind to
instead of ::. This requires a small change to pyghmi which has been
proposed upstream Ice82bf788269a31819e4495e15b9ce19a1dd057b. Until this
lands upstream pyghmi is installed from my github fork.

Other approaches were considered:
- multiple neutron ports, single openstackbmc bound to ::, this would
  require switching the pyghmi socket recv function to recvmsg to get
  the receiving IP address, but recvmsg isn't available in python 2.7
- single neutron port, single openstackbmc, modified pyghmi to select
  based on different bmc username/password. This looks impractical based
  on the assumptions the pyghmi auth code makes.
- single neutron port, multiple openstackbmc units each bound to a
  different UDP port other than 623. Impractical without many changes to
  Ironic which assumes 623.