Allow loading node details from YAML
build-nodes-json has parameters for specifying things like the node
prefixes and networks, but all of that information is already
available in the env.yaml file, so we should be able to make use of
that fact.

This did necessitate some minor changes to the quintupleo template
to remove the mostly duplicate *_net_name params so there weren't
two possible places to get the network names.  In retrospect this
probably makes more sense anyway, so although it's a backwards
incompatible change I think it's worth it.