Use a single Neutron port for bmc
The original single bmc change configured the bmc instance with
two Neutron ports.  This was done mostly to avoid some extra
manual network configuration in the bmc instance, but it resulted
in duplicate routes to the private network.  Recently it has come
to my attention that Ironic is having intermittent problems
talking to the bmc instance, which may be caused by this duplicate
route.  I had similar problems with the initial single bmc change
because it added N routes to the bmc when N baremetal instances
were created.  Reducing the duplicate routes probably mitigated
the problem, but didn't eliminate it.

This change switches to a single Neutron port for the bmc and does
the necessary configuration via os-net-config so the default route
and the Neutron assigned address will continue to work.  It also
removes the note in the documentation about needing to allow
multiple ports on a single network.