Add deploy.py CLI interface
This is intended to replace the informal bin/deploy-* scripts that
were previously used for basically the same purpose.  It combines
both scripts into one, and adds a few useful features like the
ability to define an id for the stack and have that id appended to
all resources.  Unit tests for most of the code are also added.

This change also updates the load-test-job script to use the new
CLI and renames it to be more general since it is also useful for
basic automated OVB testing.