Improve --id behavior
Previously the value passed for --id was appended to the stack
name, but this means that if you pass params like --name test --id test
then you get a stack named test-test, which probably isn't what you
wanted.

Additionally, this change sets parameter_defaults for the
network-isolation parameters so that those networks don't conflict
when the appropriate environments are in use.