Cache server status
In a large environment with a lot of bmcs running, just the periodic
status checks from something like Ironic can generate a lot of api
traffic on the host cloud.  Since nothing should be changing the
status of the instances out from under us, we can cache the status
and only look it up when we know it may have changed.  This should
reduce the number of api calls significantly.

For reference, it looks like Ironic checks power status once per
minute by default.  In an environment with 5 baremetal nodes per
baremetal stack and 50 total stacks (not a particularly huge
environment), this can generate 250 server gets per minute, which
seems to have the knock-on effect of generating a lot of Neutron
traffic too.  My assumption is that Nova is retrieving network
details for the server each time the server is retrieved, and in a
net-iso stack this involves a lot of networks and ports.