Add shared network support for quintupleo
Because of some changes we are wanting to make in
nodepool.openstack.org, we need the ability to control the shared
network setting.

We'll be now using 2 projects to configure tripleo-ci, so baremetal
nodes and zuul workers will not be in the same project any more. This
means we need to share networks between more then 1 project.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>