Support keystone v3
Make all the various bits of OVB that interact with the host cloud
able to use keystone v3 authentication.

Note that my experience has been that a newer novaclient is needed
on the bmc in order for this to work.  The pre-built BMC image in
the docs has been updated with a new enough novaclient.