Make bmc caching configurable
In ci we want to enable caching of instance status to reduce load
on the host cloud, but for interactive developer setups it can be
confusing.  If a user shuts off a baremetal vm without using the
BMC then the BMC won't recognize that its state has changed.

This change makes caching configurable, and disabled by default.
In CI we can enable it explicitly and regular users won't get any
confusing behavior by default.