Recommend using quintupleo method
Deploying a QuintupleO environment is usually going to be much
easier than starting from a base baremetal environment and adding
all of the necessary bits.  This is particularly true when using
port-security to allow undercloud DHCP because then the user is
responsible for shutting off port-security on the undercloud port.