Always use os-client-config to create openstack clients
os-client-config can handle a lot of the complexity around this that
was previously in auth.py (and the auth.py version was not entirely
correct, so it didn't work in every case).  This reduces the amount
of code that has to live in OVB considerably and should be more
robust.

The one place we continue to support directly creating clients is
in openstackbmc because there may be existing images out there that
don't have os-client-config installed and I don't want to break them.
However, that path is now deprecated and may be removed at some
point in the future.  The preferred method there is still to use
os-client-config.