Add v2 versions of net-iso templates
This creates a new directory called overcloud-templates and adds
v2 versions of the existing ipv4, ipv6, and bond net-iso templates
to it.  The existing v1 templates will also be copied to the new
directory and the old top-level directories will be deprecated.

Note that the v2 templates default to the 192.168.24.0/24 CIDR that
is the default in all releases that support v2 net-iso templates.
Otherwise they are configured identically to the v1 templates.