Support deploying with multiple Heat environments
Big changes here.  This will make it easier to deploy some of the
more advanced OVB options like network-isolation.  Instead of users
having to figure out which lines in the sample env.yaml to uncomment
when deploying advanced options, they will simply be able to include
one environment file that contains all the appropriate parameters
to turn on the option.

This _should_ be backwards compatible with existing monolithic
env.yaml configurations.  If it is not, that is a bug.

To create the option environments, a slightly modified copy of the
sample environment generator from tripleo-heat-templates has been
copied into the project.  The sample environments should _not_ be
edited directly.  Instead, the definition in the sample-env-generator
directory should be altered and the tool re-run to update the
environment.