Add test-job-v2
This is added as a completely separate script so that we can keep
testing the previous style of deployment too.  The new script uses
the option environments for configuring the OVB stack instead of a
monolithic env file.  It also tests network isolation since that is
more common and provides more test coverage of things like the
overcloud network templates too.