Fix bad assumption in build-nodes-json
A recent change assumed that the environment passed in would always
have a parameter_defaults section that contained the necessary
values.  However, since it is still possible to use an old-style
monolithic environment that doesn't use parameter_defaults that was
a bad assumption.  This fixes build-nodes-json to look at both
sections.