Catch exceptions when polling stack
In some clouds we can get intermittent failures polling the Heat
stack.  We don't really want that to fail the deployment as a whole,
so catch those exceptions and keep trying.  The exception message is
printed so the user can decide if the problem is serious enough to
stop the deployment manually.