Add boot-from-volume support
In some clouds ephemeral storage is much slower than volume storage,
so we want to be able to use volume storage for the performance-
sensitive instances.  This change adds new environments to enable
that and updates build-nodes-json to handle the case where there is
no image associated with an instance.