Sane defaults for role resource_registry
Most of the resource_registry shouldn't be used in role environments,
but for a couple of keys we should default to the same value as the
base stack.  This avoids having to duplicate network configuration
in each role, unless that role wants a different configuration.

This also adds BMCPort as an inherited key since it will almost
always need to be the same in the role stack as the base stack.