Write split node files for extra roles
There are multiple use cases for tracking the nodes of a given role
using a separate JSON file.  This change adds output files for each
role individually, in addition to the standard nodes.json that
includes everything in the deployment.  If no roles are being used,
the behavior is the same as before.