Add envup script
This is just a simple helper script that restarts the undercloud and
bmc instances for a given environment.  I shut down my local cloud
when it isn't going to be used for a while, so when I bring it back
up I have to restart all the existing instances.  This saves a little
bit of typing.