Fix default env file handling
When multiple environment support was added, backwards compatibility
was supposed to be maintained by setting the env list to env.yaml
if no envs were passed.  Unfortunately, this was done _after_ the
env list was read so it was ineffective.

This is a fairly uncommon use case as most OVB deploys will use at
least one environment file these days, but it should work so this
commit fixes it.