Prevent overwrite of input env file
When deploying an environment with --id, it is possible for the
input env file to be overwritten.  For example, consider the deploy
command `bin/deploy.py --quintupleo --id foo -e env-foo.yaml`.
Because the output id env file would be named the same as the input
env file the input env file would be overwritten and lost.

This change prevents such a situation by checking for an input env
file named 'env-ID.yaml' and raising an error if one is found.