Use xlarge flavor for undercloud nodes
Minimal requirements for undercloud nodes are described here:
https://images.rdoproject.org/docs/baremetal/requirements.html

Specifically, it says:

"The undercloud VM or baremetal machine requires: [...] 16 GB free
memory"

It's not just documentation that asks to change default flavor. If you
try to deploy using tripleo-quickstart in RDO cloud / OVB, you will hit
both tripleo-validations failures like:

There are 4 cores in the system, but there should be at least 8.
The RAM on the undercloud node is 7822 MB, the minimal recommended value
is 16384 MB.

And even when validations are ignored, the resulting environment fails
because mariadb (or another crucial service) is killed by oom-killer. I
personally experienced mariadb killed and as a result, heat-api
failing when listing stacks with ECONNREFUSED from db server, which
finally resulted in overcloud deployment failure.

xlarge is the smallest default flavor that provides 16gb ram, and so
this patch uses this flavor.