Quote delimiter to avoid heredoc expansion
Previously if a user had a shell character like ` in, for instance,
a password, it would cause the install_openstackbmc.sh script to
fail because bash would attempt to expand such characters.  Quoting
the delimiter tells bash not to expand anything in the heredoc and
just use it verbatim.  This is what we want.