Correctly retrieve disk size for boot from volume
Previously the flavor size was used unconditionally, which in the
boot from volume case would generally be incorrect.  This is
particularly problematic with flavors of a 0 disk size that can only
be used with boot from volume as the resulting Ironic nodes appear
to have no storage.

This change adds functionality to look up the volume size if we
detect that a baremetal instance was booted from volume.