Use tripleo-repos instead of rdo-release.rpm
After an upstream release, there is a period of time where the RDO
release of TripleO has not yet happened, and as a result the
rdo-release.rpm repo will not contain TripleO packages.  This breaks
the BMC because it relies on os-net-config.  Using the tripleo-repos
package to set up the RDO repos avoids this problem.