Add larger version of ipxe-boot.qcow2 image
This image can be used as the base for baremetal instances with a
flavor that has a disk size of 0, which means the image's virtual
size will be used as the base disk size.

The original image is left unchanged so it can still be used for
instances where a disk size of less than 41 GB is desired.