Remove security group from extra node port template
These extra nodes are likely to need to run arbitrary services, so
it's not ideal to have a security group that only allows port 22.
Also, the floating ip version of this template doesn't have a
security group and that one actually exposes the port on an external
network, so there's no need to lock down this one that only exposes
it to the private network.