Only add identifyer if arg.id is set
When deploying routed networks using quintupleo and roles
without '--id' arg specified the role stack fail with
property error:
 resources.internal_port.properties.network: \
   Error validating value 'overcloud_internal-None'

By only appending id if arg.id is set, the above problem is
fixed.

Additionally: The TripleO undercloud hard codes the physical
network for the provisioning network which is L2 local to the
undercloud to 'ctlplane'.

It makes sense to set the following in ovb environment:
  provision_net: ctlplane
  provision_net2: ctlplane-leaf1
  provision_net3: ctlplane-leaf2

By ommitting the '-id <id>' parameter in combination with the
above  the nodes.json build by build-nodes-json works out of
the box.