routed networks - output network environment data
Prior to routed networks the OVB workloads could use any
IP addressing, since it did not rely on any infrastructure
networking. With routed networks the workloads must use
IP addressing in the subnets in the OVB infrastructure to
enable use of the routers and dhcp_relay.

* Use allocation pool's to control the OVB infrastructure
  use of addresses in IP subnets.
* Add stack output to templates containing iformation
  about the infrastructure provisioned. I.e the addresses
  of routers in the different subnets.

Also make the dhcp_servers to which the dhcp_relay instance
will relay dhcp request to configurable.