Fix a key error in _process_role() + pep8 and py27 tests
The role does not always override the network information.
Update the networks in role_env, get the network from
parameter_defaults, fallback to parameters if not set and
finally if the netwok is not in parameters set default.

Also make default for networks: in templates json instead
of literal string.