Add router on the public network to provide external access
TripleO CI currently configures an interface on the
undercloud connected to the public network and uses the
undercloud as the router for the public network. This
deviates from what a non CI deployment would.

This change adds an optional undercloud-network-public-router
template with a router on the public_net which can provide
NAT'ed external access for overcloud nodes that use External
network interface as the default route.

The undercloud-networks-routed template have the public-router
added as well.

This removes the need for undercloud to provide masqueraded
routing for the external network when these templates are
used.