Add external interface to routed networks provision router
When deploying TripleO overcloud nodes using the ctlplane
network as the default gateway need to reach the internet
(ntp servers etc.). Previously this was done using the
undercloud as a masquerading router, doing so when nodes
are not on the same L2 network as the undercloud is not as
straight forward. (I.e we would have to set up routes on
the provision router in ovb with a default route via the
ip-address of the undercloud.)

Hooking up the router for the provision networks to the
external_net and let the ovb infra router do the NAT'ing
makes more sense.