Add symlinks for old port-security environments
For ease of migration to OVB 2.0, let's provide a compatibility
interface from the old names. A deprecation warning is also added
to deploy.py to notify users that they are using outdated
environments and should switch.

port-security.yaml is not symlinked because there's no analog in
OVB 2.0. port-security.yaml is just the default now. I suppose that
could be an empty file, but I'm not sure anyone is using it
anyway so for the moment I'm omitting it. That decision can be
revisited if necessary.